requirement.txt
mysql-connector-python
cryptography
python-dotenv